cv-back-end/requirements.txt
boto3==1.42.47
botocore==1.42.47
Django==6.0.2
django-autoslug==1.9.9
django-better-admin-arrayfield==1.4.2
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
filebrowser-safe==1.1.1
gunicorn==25.0.3 
Pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1

